+++ packages/core/src/shared/utilities/pathFind.ts
